Big fix: Env already enforces failure replacement, so that failure-replace policy is redundant.
Also, I incorrectly implemented the failure-replace (FR) policy as it should have given the same average reward as the original policy, but it didn't. Rather, it provided better reward, lol. This is because my FR policy passed incorrect number of actions to the env. It passed '[[0], [0], [0], [0]]' instead of [[0, 0], [0, 0], [0, 0], [0, 0]], i.e. no action was passed for the second segment of each edge.

Line 165-167 of env did not throw an error,
        for segment, action in zip(self.segments, actions):
            segment_reward = segment.step(action)
            reward += segment_reward
because 'zip' function ignores the extra elements in the longer list (self.segments).
for segment, action in zip([2,1], [1]):
    print(segment, action)
    >> 2 1

So, it only counted one action when computing the reward, but the replacement action was nonetheless carried out by the step function of the segment.