Add worst-case travel time calculation
Open issue:
For the Cologne-v1 env, values of self.percomputed_edge_volumes (JAX) is different from self.initial_edge_volumes (numpy). But the sum of the two arrays is matching, which implies that the TA leads to different results in each case.

This leads to a difference in worst_case_travel_times when actions are taken. The JAX env's worst case travel times is ~14% lower.

For the ToyExample-v2 env, everything works fine